image_spool/edge_od/app/requirements.txt
argparse
protobuf==3.18.0
numpy>=1.20.0
opencv-python~=4.5.3.56
tflite-support>=0.4.2
# protobuf>=3.18.0,<4 Fix by tf-team
python-dotenv
supabase